add: Also tag the latest build for alpine
It is quite common, and upstream postgres does so as well, to also tag a 'latest' for alpine. In this case, this would result in `postgis:alpine` being the latest stable alpine release. This makes it easier (albeit riskier) to track the latest stable postgis release. The risk should be acceptable however, as the same risk would be taken when tracking `postgis:latest` so it really is up to the user.